+++ cogs/snipe.py
@@ -9,7 +9,7 @@ from discord.ext import tasks
 
 import cogs.utils.carepackage as CarePackage
 import cogs.utils.rewards as Rewards
-from data import code as Database
+from data import api as Database
 
 from cogs.utils.snipe_logic import do_snipe, get_leaderboard
