+++ cogs/utils/carepackage.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from data import code as Database
+from data import api as Database
 
 
 def set_carepackage(keyword, time, hint):
